requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.0
google-generativeai==0.3.2
python-json-logger==2.0.7
structlog==24.1.0
ortools==9.8.3296